+++ src/main/resources/db/migration/V13__create_status.sql
@@ -1,18 +1,18 @@
 CREATE TABLE status (
-                        id SERIAL PRIMARY KEY,
-                        nome VARCHAR(100) NOT NULL,
-                        codigo VARCHAR(50) NOT NULL,
-                        descricao TEXT,
-                        ordem INTEGER,
-                        ativo BOOLEAN DEFAULT TRUE,
+    id BIGSERIAL PRIMARY KEY,
+    nome VARCHAR(100) NOT NULL,
+    codigo VARCHAR(50) NOT NULL,
+    descricao TEXT,
+    ordem INTEGER,
+    ativo BOOLEAN DEFAULT TRUE,
 
-                        tipo_status_id INTEGER NOT NULL,
+    tipo_status_id BIGINT NOT NULL,
 
-                        created_at TIMESTAMP,
-                        updated_at TIMESTAMP,
-                        deleted_at TIMESTAMP,
+    created_at TIMESTAMP,
+    updated_at TIMESTAMP,
+    deleted_at TIMESTAMP,
 
-                        CONSTRAINT fk_tipo_status
-                            FOREIGN KEY (tipo_status_id)
-                                REFERENCES tipo_status(id)
+    CONSTRAINT fk_tipo_status
+        FOREIGN KEY (tipo_status_id)
+            REFERENCES tipo_status (id)
 );
